src/webapp/requirements/base.txt
django==5.2.0
django-allauth[socialaccount]==65.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn=23.0.0
psycopg[binary]==3.2.6
